Do it as event
