+++ sources/Adapters/adv/gui/advGUIWindowImp.cpp
@@ -11,11 +11,15 @@
 #ifdef USB_REMOTE_UI
 #include "picoRemoteUI.h"
 #endif
+#include "ff_gen_drv.h"
+#include "sd_diskio.h"
 #include <string>
 
 #define to_rgb565(color)                                                       \
   ((color._r & 0b11111000) << 8) | ((color._g & 0b11111100) << 3) |            \
       (color._b >> 3)
+
+extern char SDPath[4]; /* SD logical drive path */
 
 // classic picotracker mapping
 static GUIEventPadButtonType eventMappingPico[11] = {
@@ -188,6 +192,14 @@ void advGUIWindowImp::ProcessEvent(advEvent &event) {
   case PICO_CLOCK:
     instance_->_window->ClockTick();
     break;
+  case PICO_SD_DET:
+    // SD reinit
+    FATFS_UnLinkDriver(SDPath);
+    HAL_SD_DeInit(&hsd1);
+    __HAL_RCC_SDMMC1_FORCE_RESET();
+    __HAL_RCC_SDMMC1_RELEASE_RESET();
+    MX_SDMMC1_SD_Init();
+    FATFS_LinkDriver(&SD_DMA_Driver, SDPath);
   case LAST:
     break;
   }
